+++ init_submodules.sh
@@ -7,7 +7,7 @@ home=$(pwd)
 git submodule sync
 
 git submodule update --init benchmarks/coremark
-
+git submodule update --init benchmarks/mibench2
 git submodule update --init riscv-openocd
 
 git submodule update --init riscv-tests
@@ -15,11 +15,24 @@ cd riscv-tests
 git submodule sync
 cd ..
 
-git submodule update --init FreeRTOS-mirror busybox newlib
+git submodule update --init FreeRTOS-mirror
+git submodule update --init busybox
+git submodule update --init newlib
 git submodule update --init --recursive \
-bluespec-processors/P1/Piccolo bluespec-processors/P2/Flute \
-riscv-linux bluespec-processors/P3/Toooba riscv-tests riscv-pk
+bluespec-processors/P1/Piccolo
+git submodule update --init --recursive \
+bluespec-processors/P2/Flute
+git submodule update --init --recursive \
+bluespec-processors/P3/Toooba
+git submodule update --init --recursive \
+riscv-linux
+git submodule update --init --recursive \
+riscv-tests
+git submodule update --init --recursive \
+riscv-pk
 git submodule update --init chisel_processors
+git submodule update --init freebsd/cheribsd
+
 
 cd chisel_processors
 git submodule sync