+++ install/deps.sh
@@ -3,29 +3,37 @@
 # It should only be run as root, once per host, from the repo root dir.
 set -eux
 
+# General toolchain dependencies
+apt-get install -y automake autotools-dev libmpc-dev libmpfr-dev libgmp-dev gawk
+apt-get install -y build-essential gperf patchutils zlib1g-dev libexpat-dev
+
 # For riscv-linux build:
-apt install -y curl openssl bc bison flex make autoconf debootstrap proot libssl-dev debian-ports-archive-keyring
+apt-get install -y curl openssl bc bison flex make autoconf debootstrap proot
+apt-get install -y libssl-dev debian-ports-archive-keyring device-tree-compiler
 
 # RTL simulator and RISC-V emulator:
-apt install -y verilator qemu qemu-user qemu-system-misc
+apt-get install -y verilator qemu qemu-user qemu-system-misc
 
 # Needed for GDB
-apt install -y libpython2.7
+apt-get install -y libpython2.7
+
+# OpenOCD dependencies
+apt-get install -y libftdi1-2 libusb-1.0-0-dev
 
 # Needed for manual tests
-apt install -y minicom
+apt-get install -y minicom
 
 # System-wide python packages needed by testing scripts
-apt install -y python3-pip
+apt-get install -y python3-pip
 pip3 install pyserial pexpect
 
 # Xilinx vivado_lab dependency
-apt install -y libtinfo5
+apt-get install -y libtinfo5
 
-# Clang and LLVM for RISC-V:
-wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | apt-key add -
-add-apt-repository 'deb http://apt.llvm.org/buster/ llvm-toolchain-buster main'
-apt-get update
-apt-get install -y clang-10 lldb-10 lld-10 clangd-10
-# TODO: set up a symlink
-ln -s /usr/bin/clang-10 /usr/bin/clang
+# Instal dependencies for FreeBSD
+apt-get install -y libtool pkg-config cmake ninja-build samba texinfo libarchive-dev
+apt-get install -y libglib2.0-dev libpixman-1-dev libarchive-dev bsdtar libbz2-dev
+
+# Install and configure TFTP server
+apt-get install -y atftpd
+chmod 777 /srv/tftp