+++ setup_env.sh
@@ -34,8 +34,8 @@ function check_file {
 }
 
 function proc_usage {
-    echo "Usage: $0 [chisel_p1|chisel_p2|chisel_p3|bluespec_p1|bluespec_p2|bluespec_p3]"
-    echo "Please specify a bluespec or chisel processor!"
+    echo "Usage: $0 <chisel_p1|chisel_p2|chisel_p2_pcie|chisel_p3|bluespec_p1|bluespec_p2|bluespec_p2_pcie|bluespec_p3> [secure-boot-binary-image]"
+    echo "Please specify a bluespec or chisel processor, and optionally a binary image"
 }
 
 function proc_picker {
@@ -44,16 +44,21 @@ function proc_picker {
 	        proc_name="bluespec_p1"
 	elif [ "$1" == "bluespec_p2" ]; then
 	        proc_name="bluespec_p2"
+	elif [ "$1" == "bluespec_p2_pcie" ]; then
+	        proc_name="bluespec_p2_pcie"
 	elif [ "$1" == "bluespec_p3" ]; then
 	        proc_name="bluespec_p3"
 	elif [ "$1" == "chisel_p1" ]; then
 	        proc_name="chisel_p1"
 	elif [ "$1" == "chisel_p2" ]; then
 	        proc_name="chisel_p2"
+	elif [ "$1" == "chisel_p2_pcie" ]; then
+	        proc_name="chisel_p2_pcie"
 	elif [ "$1" == "chisel_p3" ]; then
 	        proc_name="chisel_p3"
 	else
 	        proc_usage
+            echo "Please specify a bluespec or chisel processor!"
 	        exit -1
 	fi
 }